refactor: change 'AssetType' rawValue type to Int16
- 코어데이터 typeValue(Int16)과의 호환성

+++ AccountBook/Models/AssetItemManager.swift
@@ -7,7 +7,7 @@
 
 import Foundation
 
-enum AssetType: Int, CaseIterable {
+enum AssetType: Int16, CaseIterable {
     case cash = 0
     case bankAccount = 1
     case debitCard = 2
